// packages/design-system/src/components/text/text.enum.ts
export enum TextVariant {
  // Heading
  HeadingPrimary = "HeadingPrimary",
  HeadingSecondary = "HeadingSecondary",
  HeadingTertiary = "HeadingTertiary",
  HeadingQuaternary = "HeadingQuaternary",

  // Large texts
  Large = "Large",
  LargeBold = "LargeBold",

  // Basic
  Basic = "Basic",
  BasicBold = "BasicBold",
  BasicGrey = "BasicGrey",

  // Small
  Small = "Small",
  SmallBold = "SmallBold",
  SmallGrey = "SmallGrey",
  SmallLight = "SmallLight",
  SmallLightGrey = "SmallLightGrey",

  // Tiny
  Tiny = "Tiny",
  TinyGrey = "TinyGrey",
  TinyBold = "TinyBold",
  TinyBoldGrey = "TinyBoldGrey",
}

export enum TextDecorator {
  Accent = "Accent",
  Error = "Error",
  UpperCase = "UpperCase",
  Fade = "Fade",
  Invert = "Invert",
}